Fix: correct syntax needed to return newly-inserted data
When a user first logs in, a new User instance is created in the
database. This newly-created user data was not properly being returned
to the controller however, resulting in a failure to create a JWT and so
on.

This fix involves using the correct Knex.js syntax to return data after an
`INSERT`.